+++ client/ui/instruments/__init__.py
@@ -3,7 +3,9 @@
 from client.ui.instruments.base import (POLL_KINDS, POLL_MEASUREMENTS,
                                         POLL_READINGS, POLL_STATE,
                                         InstrumentPanel)
+from client.ui.instruments.electronic_load import ElectronicLoadPanel
 from client.ui.instruments.generic import GenericInstrumentPanel
+from client.ui.instruments.multimeter import MultimeterPanel
 from client.ui.instruments.oscilloscope import OscilloscopePanel
 from client.ui.instruments.power_supply import PowerSupplyPanel
 from client.ui.instruments.registry import (panel_class_for, register_panel,
@@ -16,6 +18,8 @@ __all__ = [
     "GenericInstrumentPanel",
     "PowerSupplyPanel",
     "OscilloscopePanel",
+    "ElectronicLoadPanel",
+    "MultimeterPanel",
     "POLL_KINDS",
     "POLL_READINGS",
     "POLL_MEASUREMENTS",